requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
mlflow>=2.0.0
lightgbm>=3.3.0
xgboost>=1.5.0
catboost>=1.0.0
optuna>=2.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
joblib>=1.0.0